posts/rl_cpt_value/requirements.txt
gymnasium
numpy
Pillow
openai
stable-baselines3
python-dotenv
